[PB-570](fix): Fix PR comments
